Created basic infofile generator for testing cctaskrunner.ini logging path fix improved kill.sh looping in infofile listing in startcc.sh

+++ kill.sh
@@ -1,13 +1,44 @@
 #! /bin/sh
 
+set -e
+
 # requires -9 for some webscale reasons...
-killall -9 mongrel2
 
-pkill -f hots
-pkill -f ccserver
-pkill -f cctaskrunner
-pkill -f infosender
+if ps -ef | grep -v grep | grep mongrel2
+then
+    echo "killing mongrel2"
+    killall -9 mongrel2
+fi
 
-rm -rf ./var/log ./var/run ./var/info*
+if [ -f var/run/hots.pid ]
+then
+    echo "killing hots"
+    kill -9 `cat var/run/hots.pid`
+fi
+
+if [ -f var/run/cclocal.pid ]
+then
+    echo "killing cclocal"
+    kill -9 `cat var/run/cclocal.pid`
+fi
+
+if [ -f var/run/ccserver.pid ]
+then
+    echo "killing ccserver"
+    kill -9 `cat var/run/ccserver.pid`
+fi
+
+if [ -f var/run/cctaskrunner.pid ]
+then
+    echo "killing cctaskrunner"
+    kill -9 `cat var/run/cctaskrunner.pid`
+fi
+
+if [ -f var/run/infosender.pid ]
+then
+    echo "killing infosender"
+    kill -9 `cat var/run/infosender.pid`
+fi
+
+rm -rf ./var/log ./var/run ./var/info* ./infodir/info*
 mkdir -p var/log var/run
-
